SIH2022-Project-main/requirements.txt
tkinter
customtkinter
Pillow
opencv-python
face_recognition
fpdf
